new rebuild support

+++ scripts/scudum/util/make.img.sh
@@ -2,13 +2,19 @@ FILE=${FILE-scudum.img}
 DEV_NAME=${DEV_NAME-/dev/null}
 BOOT_SIZE=${BOOT_SIZE-+1G}
 SWAP_SIZE=${SWAP_SIZE-+2G}
+REBUILD=${REBUILD-0}
 
 DIR=$(dirname $(readlink -f $0))
 
-dd if=/dev/zero of=$DEV_NAME bs=1M
+if [ "$REBUILD" == "1" ]; then
+    dd if=/dev/zero of=$DEV_NAME bs=1M
 
-DEV_NAME=$DEV_NAME BOOT_SIZE=$BOOT_SIZE\
-    SWAP_SIZE=$SWAP_SIZE $DIR/install.dev.sh
+    DEV_NAME=$DEV_NAME BOOT_SIZE=$BOOT_SIZE\
+        SWAP_SIZE=$SWAP_SIZE $DIR/install.dev.sh
+fi
 
 dd if=$DEV_NAME of=$FILE bs=1M
-dd if=/dev/zero of=$DEV_NAME bs=1M
+
+if [ "$REBUILD" == "1" ]; then
+    dd if=/dev/zero of=$DEV_NAME bs=1M
+fi
